Adds hacky singleton service for checking roles.
Firebase Auth, as of May 1st 2020, is one of the most unpleasant pieces
of software that I've had the chance to work with. There's no good way
to store role information so you're forced to jam role information into
their custom claims which is more like a hot pile of unusable mess as
there's no easy way to check custom claims since it comes with
getIDTokenResult. (Who even came up with that name?!)

Anyways, this commit adds in a hacky singleton service to imitate a
Redux Store because relying on the Firebase Auth state change listener
was too slow and caused a lot of bugs. Hopefully no one has to ever read
this code again but if you do I'm terribly sorry.