Added base AuthMiddleware and OfficerAuth and MemberAuth.
AuthMiddleware works very similarly to the frontend RBAC HOCs.
It's just that for AuthMiddleware, we use Firebase Admin SDK
instead of React Context.

OfficerAuth uses AuthMiddleware to be an Auth middleware that
checks if a user is an officer.

MemberAuth uses AuthMiddleware to be an Auth middleware that
checks if a user is a member.