Decoupled verifyRole() from AuthMiddleware even more.
verifyRole() now only takes a token string and array of permitted roles
(used to take the 'next' function as well). verifyRole() now returns a
boolean, true to indicate that the user is authenticated and has a
permitted role, and false to indicate that the user is authenticated
and does not have any of the permitted roles. verifyRole() throws an
error message if admin.auth().verifyIdToken() fails (unauthenticated
ID token).

AuthMiddleware now has a try catch block as well, so if verifyRole()
throws an error, called 'err' in this case, AuthMiddleware can call
next(err) to follow the Express error handling convention. If
verifyRole() runs fine, then it will return a boolean, indicating
whether the authenticated user has the necessary roles for the action
they want to do.