Added joi middleware to provide a good way of validating req.body.
Since we have decided to use joi as the main library for validating
req.body (request body) in our back end code, I wrote this middleware
to make that task a lot less painful and to modularize the code better.

To use the middleware, import from the src/middlewares/validation/base
folder. Then, use it as a middleware on HTTP requests that require
the use of req.body.

Example:

app.post('/', ValidationMiddleware(schema), (req, res) => {
  res.json({
    Success: 'Success!'
  });
});