Made changes to email service so users won't need to choose a sender.
Created a separate type in email.definition.ts called EmailInfoRequired
that combines EmailInfo
and 'from' property to conform to MailDataRequired type of SendGrid's
send().
Made a function that inserts 'from' fields into EmailInfo objects and
returns them as EmailInfoRequired objects

Added a public method (sendEmail()) that sends an email to one or more
people.

Note that in the EmailInfo interface, 'from' property is still defined,
but is now optional.