Changed BaseModal and created ButtonWithModal.
BaseModal's opening and closing now relies on the useState hook
in ButtonWithModal. ButtonWithModal is just a modal with a button
that can be used to open that modal.

In addition to that, ConfirmationModalWithButton and
AlertModalWithButton are also created using ButtonWithModal.
ConfirmationModalWithButton is basically a two-choice modal with
a button to open it. AlertModalWithButton is a modal with only
one action.