Changed all pages that didn't use UserContext to using them.
PointsPage now use UserContext to check if a user is a member (aka
member or officer) or not. InducteePointsPage now doesn't even
check for roles, since routing HOC limited access to just officers
already.

Fixed up InducteePointsPage so that it works now because the db
migration broke a lot of the service functions that we have.