Peel back controller complexity.
Heads up this is a fat commit because I didn't commit as much as I
should have.

Services are now directly injected via @inject(Service) instead of
registering them with an interface. While DI against an interface is the
"right" thing to do, we treat our services as a unit of business logic,
and thus there is no need to swap them out with DI interfaces. This
makes development a little easier as there's no need to go through the
ritual of writing interfaces in order to develop a single use business
logic service.

For the use cases of swapping out email/S3/auth services, we can do that on a
case by case basis and onboard the interface complexity as needed.

Services/Controllers are now @singleton instead of @injectable. Since
they are all stateless bags of functions, we can share a single instance
globally.

Repositories are now injected properly so we can test our services a
little better. without having to resolve to using jest.mock()/jest.fn().
Unfortunately due to the nature of typeorm's getRepository call, we are
forced to register our repositories using a token and resolving it via
said tokens instead of being able to do something like @injectRepository.
All repositories and their factories are in repositories/index.ts.
As a result, repositories now have to be loaded in
loaders/RepositoryLoader.ts.

Services have been stripped of their dependencies on payloads. They now
directly operate on entities. Entity to payload conversion and vice
versa has been offloaded to a new type of class called mappers. All
these mappers do is translate payloads into entities, and entities into
payloads.