Add CardWithVerticalTabs, with 3 tables to check off, view attendances, and view RSVPs (#211)
The Attendance/RSVP/Checkoff table is ready to be released. Most of the CSS stuff is fixed except for one where when the browser is at half-screen, the table goes under the tabs themselves. We plan to eventually address this, but for now we'll leave it there for later. Attached is a picture of the CSS bug.